micros/quotes-api/requirements.txt
deta
fastapi
python-dotenv